Add - added a  linear interpolator for the reference velocity table

+++ Code/airbrakes/controll.h
@@ -12,7 +12,7 @@ typedef struct Parameters_t {
 // reference height, velocity pair. Height is index and 
 extern const float reference[3300];
 
-float controller(Rocket* rocket, float* error, Parameters* parameters, float* riemann_sum, float dt, unsigned int height, float velocity);
+float controller(Rocket* rocket, float* error, Parameters* parameters, float* riemann_sum, float dt, float height, float velocity);
 float integrate(float prev_sum, float value, float step);
 
 #endif
